added dtype for layers and optimizer

+++ optimizers.py
@@ -10,17 +10,17 @@ class Adam:
         self.fme = {}  # First moment estimate
         self.sme = {}  # Second moment estimate
 
-    def adjust_gradient(self, layer, weight_gradient, bias_gradient, learning_rate):
+    def adjust_gradient(self, layer, weight_gradient, bias_gradient, learning_rate, dtype=np.float32):
         if layer not in self.fme:
             # Initialize moments for this layer
-            self.fme[layer] = [np.zeros_like(weight_gradient), np.zeros_like(bias_gradient)]
-            self.sme[layer] = [np.zeros_like(weight_gradient), np.zeros_like(bias_gradient)]
+            self.fme[layer] = [np.zeros_like(weight_gradient, dtype=dtype), np.zeros_like(bias_gradient, dtype=dtype)]
+            self.sme[layer] = [np.zeros_like(weight_gradient, dtype=dtype), np.zeros_like(bias_gradient, dtype=dtype)]
 
         # Update moments first
-        self.fme[layer][0] = self.b1 * self.fme[layer][0] + (1 - self.b1) * weight_gradient
-        self.fme[layer][1] = self.b1 * self.fme[layer][1] + (1 - self.b1) * bias_gradient
-        self.sme[layer][0] = self.b2 * self.sme[layer][0] + (1 - self.b2) * np.square(weight_gradient)
-        self.sme[layer][1] = self.b2 * self.sme[layer][1] + (1 - self.b2) * np.square(bias_gradient)
+        self.fme[layer][0] = (self.b1 * self.fme[layer][0] + (1 - self.b1) * weight_gradient).astype(dtype)
+        self.fme[layer][1] = (self.b1 * self.fme[layer][1] + (1 - self.b1) * bias_gradient).astype(dtype)
+        self.sme[layer][0] = (self.b2 * self.sme[layer][0] + (1 - self.b2) * np.square(weight_gradient)).astype(dtype)
+        self.sme[layer][1] = (self.b2 * self.sme[layer][1] + (1 - self.b2) * np.square(bias_gradient)).astype(dtype)
 
         # Bias correction
         bc_m_weights = self.fme[layer][0] / (1 - self.b1 ** self.step)
@@ -32,7 +32,7 @@ class Adam:
         weight_change = learning_rate * bc_m_weights / (np.sqrt(bc_v_weights + self.epsilon))
         bias_change = learning_rate * bc_m_bias / (np.sqrt(bc_v_bias + self.epsilon))
 
-        return weight_change, bias_change
+        return weight_change.astype(dtype), bias_change.astype(dtype)
 
 
 class NoOptimizer:
